# specs/001-echo-server/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 에코 서버 API
  description: |
    WebSocket과 REST API를 지원하는 에코 서버입니다.
    개발자가 API를 테스트하고 디버깅할 수 있도록 10개의 REST API 엔드포인트와 실시간 모니터링 기능을 제공합니다.
  version: 1.0.0
  contact:
    name: Studio RainShelter
servers:
  - url: http://localhost:3000
    description: 로컬 개발 서버
  - url: http://localhost:3000
    description: 프로덕션 서버 (TBD)

tags:
  - name: Echo
    description: 에코 엔드포인트 (요청 데이터 반환)
  - name: Mock
    description: Mock 데이터 엔드포인트
  - name: Testing
    description: 테스트용 특수 엔드포인트 (지연, 에러)
  - name: Monitoring
    description: 서버 상태 및 로그 조회

paths:
  /api/echo:
    get:
      tags:
        - Echo
      summary: 쿼리 파라미터 에코
      description: 쿼리 파라미터를 JSON 형식으로 반환합니다.
      operationId: getEcho
      parameters:
        - name: message
          in: query
          description: 에코할 메시지
          required: false
          schema:
            type: string
            example: hello
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  echo:
                    type: object
                    properties:
                      query:
                        type: object
                        additionalProperties: true
                  timestamp:
                    type: string
                    format: date-time
              example:
                echo:
                  query:
                    message: hello
                timestamp: "2025-11-02T12:34:56.789Z"

    post:
      tags:
        - Echo
      summary: Body 데이터 에코
      description: 요청 Body의 JSON 데이터를 그대로 반환합니다.
      operationId: postEcho
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            example:
              data: test
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  echo:
                    type: object
                    properties:
                      body:
                        type: object
                        additionalProperties: true
                  timestamp:
                    type: string
                    format: date-time
              example:
                echo:
                  body:
                    data: test
                timestamp: "2025-11-02T12:34:56.789Z"

  /api/users:
    get:
      tags:
        - Mock
      summary: Mock 사용자 목록 조회
      description: 테스트용 Mock 사용자 목록을 반환합니다.
      operationId: getUsers
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
              example:
                users:
                  - id: 1
                    name: Alice
                    email: alice@example.com
                  - id: 2
                    name: Bob
                    email: bob@example.com

    post:
      tags:
        - Mock
      summary: 사용자 생성 시뮬레이션
      description: 사용자 생성을 시뮬레이션하고 입력 데이터를 반환합니다.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
              example:
                created: true
                user:
                  id: 3
                  name: Charlie
                  email: charlie@example.com

  /api/status:
    get:
      tags:
        - Monitoring
      summary: 서버 상태 확인
      description: 서버의 현재 상태와 통계를 반환합니다.
      operationId: getStatus
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerStatus'

  /api/json:
    post:
      tags:
        - Echo
      summary: JSON 데이터 처리
      description: 복잡한 JSON 데이터를 처리하고 반환합니다.
      operationId: processJson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  processed:
                    type: boolean
                  data:
                    type: object
                    additionalProperties: true
                  timestamp:
                    type: string
                    format: date-time

  /api/delay/{ms}:
    get:
      tags:
        - Testing
      summary: 지연된 응답 테스트
      description: 지정된 밀리초만큼 지연 후 응답합니다.
      operationId: delayedResponse
      parameters:
        - name: ms
          in: path
          required: true
          description: 지연 시간 (밀리초)
          schema:
            type: integer
            minimum: 0
            maximum: 10000
            example: 2000
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  delayed:
                    type: boolean
                  duration:
                    type: integer
                  timestamp:
                    type: string
                    format: date-time
              example:
                delayed: true
                duration: 2000
                timestamp: "2025-11-02T12:34:58.789Z"
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/update/{id}:
    put:
      tags:
        - Mock
      summary: 업데이트 시뮬레이션
      description: 리소스 업데이트를 시뮬레이션합니다.
      operationId: updateResource
      parameters:
        - name: id
          in: path
          required: true
          description: 리소스 ID
          schema:
            type: string
            example: "123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: boolean
                  id:
                    type: string
                  data:
                    type: object
                    additionalProperties: true

  /api/delete/{id}:
    delete:
      tags:
        - Mock
      summary: 삭제 시뮬레이션
      description: 리소스 삭제를 시뮬레이션합니다.
      operationId: deleteResource
      parameters:
        - name: id
          in: path
          required: true
          description: 리소스 ID
          schema:
            type: string
            example: "123"
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                  id:
                    type: string
              example:
                deleted: true
                id: "123"

  /api/error/{code}:
    get:
      tags:
        - Testing
      summary: 에러 응답 테스트
      description: 지정된 HTTP 상태 코드로 에러 응답을 반환합니다.
      operationId: errorResponse
      parameters:
        - name: code
          in: path
          required: true
          description: HTTP 상태 코드
          schema:
            type: integer
            minimum: 400
            maximum: 599
            example: 404
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          description: 요청된 상태 코드
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/logs:
    get:
      tags:
        - Monitoring
      summary: 로그 조회
      description: 저장된 요청/응답 로그를 조회합니다.
      operationId: getLogs
      parameters:
        - name: limit
          in: query
          description: 조회할 로그 수
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/LogEntry'
                  total:
                    type: integer

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Alice
        email:
          type: string
          format: email
          example: alice@example.com
      required:
        - id
        - name
        - email

    UserInput:
      type: object
      properties:
        name:
          type: string
          example: Charlie
        email:
          type: string
          format: email
          example: charlie@example.com
      required:
        - name
        - email

    ServerStatus:
      type: object
      properties:
        status:
          type: string
          example: running
        uptime:
          type: integer
          description: 서버 가동 시간 (초)
          example: 86400
        statistics:
          type: object
          properties:
            totalRequests:
              type: integer
              example: 1523
            successfulRequests:
              type: integer
              example: 1489
            failedRequests:
              type: integer
              example: 34
            activeWebSockets:
              type: integer
              example: 12
            totalWebSocketMessages:
              type: integer
              example: 3456
        timestamp:
          type: string
          format: date-time

    LogEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        request:
          type: object
          properties:
            method:
              type: string
            path:
              type: string
            query:
              type: object
            body:
              type: object
        response:
          type: object
          properties:
            statusCode:
              type: integer
            body:
              type: object
            duration:
              type: integer

    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: integer
              example: 404
            message:
              type: string
              example: Resource not found
            timestamp:
              type: string
              format: date-time
      required:
        - error

  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: 400
              message: Bad Request
              timestamp: "2025-11-02T12:34:56.789Z"

    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: 404
              message: Resource not found
              timestamp: "2025-11-02T12:34:56.789Z"

    InternalServerError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: 500
              message: Internal Server Error
              timestamp: "2025-11-02T12:34:56.789Z"